Using xdotool to verify pid for window id.
It seems window ids returned by wmctrl -l are not unique for the session, resulting toggling a different window if the original is closed, and the id is reused. Now the window id and pid must match, so if there is a different pid, the window id is considered reused and will be dropped.